Website Layout Decided On
A lot of my session today was spend decided on how the layout should be handled like, luckily I was do jus that and am taking heavy inspiration from the second project. While the Index didn't have too much progress, I was able to get a load of progress accomplished in the "br_styles.css" stylesheet.