Project Finished
While I'm not sure if it is exactly complete, I can say for sure that I am decently satisfied with how it turned out as I was able to get all the things I saw from the second project in it. I do think it looks a little bare and there was a lot of CSS work needed, but I'm glad this turned out the way it did.